+++ packages/api/src/interfaces/cli/Scdl.cli.test.ts
@@ -8,7 +8,8 @@ jest.mock("csv-stringify/sync", () => ({
 
 jest.mock("fs");
 const mockedFs = jest.mocked(fs);
-import FormatDateError from "../../shared/errors/cliErrors/FormatDateError";
+import * as CliHelper from "../../shared/helpers/CliHelper";
+jest.mock("../../shared/helpers/CliHelper");
 import ScdlCli from "./Scdl.cli";
 import scdlService from "../../modules/providers/scdl/scdl.service";
 
@@ -16,12 +17,8 @@ jest.mock("../../modules/providers/scdl/scdl.service");
 const mockedScdlService = jest.mocked(scdlService);
 import MiscScdlGrant from "../../modules/providers/scdl/__fixtures__/MiscScdlGrant";
 import { DuplicateIndexError } from "../../shared/errors/dbError/DuplicateIndexError";
-import ScdlGrantParser from "../../modules/providers/scdl/scdl.grant.parser";
 import MiscScdlProducer from "../../modules/providers/scdl/__fixtures__/MiscScdlProducer";
 import { ParsedDataWithProblem } from "../../modules/providers/scdl/@types/Validation";
-
-jest.mock("../../modules/providers/scdl/scdl.grant.parser");
-const mockedScdlGrantParser = jest.mocked(ScdlGrantParser);
 
 import csvSyncStringifier = require("csv-stringify/sync");
 import dataLogService from "../../modules/data-log/dataLog.service";
@@ -38,7 +35,6 @@ describe("ScdlCli", () => {
     const STORABLE_DATA = { ...GRANT, __data__: {} };
     const FILE_CONTENT = "FILE_CONTENT";
     const EXPORT_DATE_STR = "2023-03-23";
-    const EXPORT_DATE = new Date(EXPORT_DATE_STR);
     const UNIQUE_ID = "UNIQUE_ID";
     const FILE_PATH = "FILE_PATH";
     const STORABLE_DATA_ARRAY = [STORABLE_DATA];
@@ -54,8 +50,8 @@ describe("ScdlCli", () => {
         mockedScdlService.getProducer.mockResolvedValue(PRODUCER_ENTITY);
         // @ts-expect-error: private method
         mockedScdlService._buildGrantUniqueId.mockReturnValue(UNIQUE_ID);
-        mockedScdlGrantParser.parseCsv.mockReturnValue({ entities: STORABLE_DATA_ARRAY, errors: [] });
-        mockedScdlGrantParser.parseExcel.mockReturnValue({ entities: STORABLE_DATA_ARRAY, errors: [] });
+        mockedScdlService.parseCsv.mockReturnValue({ entities: STORABLE_DATA_ARRAY, errors: [] });
+        mockedScdlService.parseXls.mockReturnValue({ entities: STORABLE_DATA_ARRAY, errors: [] });
         cli = new ScdlCli();
     });
 
@@ -114,15 +110,15 @@ describe("ScdlCli", () => {
     }
 
     describe.each`
-        methodName    | test            | parserMethod                  | parserArgs
-        ${"parse"}    | ${testParseCsv} | ${ScdlGrantParser.parseCsv}   | ${[FILE_CONTENT, DELIMETER, QUOTE]}
-        ${"parseXls"} | ${testParseXls} | ${ScdlGrantParser.parseExcel} | ${[FILE_CONTENT, PAGE_NAME, ROW_OFFSET]}
+        methodName    | test            | parserMethod            | parserArgs
+        ${"parse"}    | ${testParseCsv} | ${scdlService.parseCsv} | ${[FILE_CONTENT, DELIMETER, QUOTE]}
+        ${"parseXls"} | ${testParseXls} | ${scdlService.parseXls} | ${[FILE_CONTENT, PAGE_NAME, ROW_OFFSET]}
     `("$methodName", ({ test, parserMethod, parserArgs }) => {
         it("sanitizes input", async () => {
             // @ts-expect-error -- test private
             const sanitizeSpy = jest.spyOn(cli, "validateGenericInput");
             await test();
-            expect(sanitizeSpy).toHaveBeenCalledWith(FILE_PATH, PRODUCER_ENTITY.slug, EXPORT_DATE_STR);
+            expect(sanitizeSpy).toHaveBeenCalledWith(PRODUCER_ENTITY.slug, EXPORT_DATE_STR);
         });
 
         it("reads file", async () => {
@@ -140,7 +136,7 @@ describe("ScdlCli", () => {
             const persistSpy = jest.spyOn(cli, "persistEntities").mockReturnValueOnce(Promise.resolve());
             jest.mocked(parserMethod).mockReturnValueOnce({ entities: STORABLE_DATA_ARRAY });
             await test();
-            expect(persistSpy).toHaveBeenCalledWith(STORABLE_DATA_ARRAY, PRODUCER_ENTITY.slug, EXPORT_DATE_STR);
+            expect(persistSpy).toHaveBeenCalledWith(STORABLE_DATA_ARRAY, PRODUCER_ENTITY.slug);
         });
 
         it("exports errors", async () => {
@@ -154,7 +150,11 @@ describe("ScdlCli", () => {
 
         it("logs import", async () => {
             await test();
-            expect(dataLogService.addLog).toHaveBeenCalledWith(PRODUCER_ENTITY.slug, EXPORT_DATE, FILE_PATH);
+            expect(dataLogService.addLog).toHaveBeenCalledWith(
+                PRODUCER_ENTITY.slug,
+                FILE_PATH,
+                new Date(EXPORT_DATE_STR),
+            );
         });
     });
 
@@ -182,24 +182,34 @@ describe("ScdlCli", () => {
         });
 
         it("should call scdlService.updateProducer()", async () => {
+            const now = new Date();
+            jest.useFakeTimers().setSystemTime(now);
             // @ts-expect-error -- test private
-            await cli.persistEntities(STORABLE_DATA_ARRAY, PRODUCER_ENTITY.slug, EXPORT_DATE_STR);
+            await cli.persistEntities(STORABLE_DATA_ARRAY, PRODUCER_ENTITY.slug);
             expect(scdlService.updateProducer).toHaveBeenCalledWith(PRODUCER_ENTITY.slug, {
-                lastUpdate: new Date(EXPORT_DATE_STR),
+                lastUpdate: now,
             });
+            jest.useRealTimers();
         });
     });
 
     describe("validateGenericInput", () => {
-        it("should throw FormatDateError", async () => {
+        it("validates export date", async () => {
             // @ts-expect-error -- test private
-            expect(() => cli.validateGenericInput(PRODUCER_ENTITY.slug)).rejects.toThrowError(FormatDateError);
+            cli.validateGenericInput(PRODUCER_ENTITY.slug, EXPORT_DATE_STR);
+            expect(CliHelper.validateDate).toHaveBeenCalledWith(EXPORT_DATE_STR);
+        });
+
+        it("does not validates date if undefined", async () => {
+            // @ts-expect-error -- test private
+            cli.validateGenericInput(PRODUCER_ENTITY.slug);
+            expect(CliHelper.validateDate).not.toHaveBeenCalled();
         });
 
         it("should throw Error when providerId does not match any provider in database", async () => {
             mockedScdlService.getProducer.mockResolvedValue(null);
             // @ts-expect-error -- test private
-            expect(() => cli.validateGenericInput("WRONG_ID", new Date())).rejects.toThrowError();
+            expect(() => cli.validateGenericInput("WRONG_ID")).rejects.toThrowError();
         });
     });
 
@@ -208,7 +218,7 @@ describe("ScdlCli", () => {
         const FILE = "path/file.csv";
         const STR_CONTENT = "azertyuiop";
         // normalize for windows and linux compatilibity
-        const OUTPUT_PATH = normalize("importErrors/file.csv-errors.csv");
+        const OUTPUT_PATH = normalize("import-errors/file.csv-errors.csv");
 
         beforeAll(() => {
             jest.mocked(csvSyncStringifier.stringify).mockReturnValue(STR_CONTENT);