+++ packages/api/src/modules/providers/dataBretagne/dataBretagne.service.ts
@@ -38,7 +38,7 @@ class DataBretagneService extends ProviderCore {
         // do not replace programs if empty
         if (!programs || !programs.length) throw new Error("Unhandled error from API Data Bretagne");
         await stateBudgetProgramPort.replace(programs);
-        await dataLogService.addLog(dataBretagneService.provider.id, new Date(), "api");
+        await dataLogService.addLog(dataBretagneService.provider.id, "api", new Date());
     }
 
     /**